[MC-1509]rename module directory to json_module directory
