+++ app/json_module/NameRecord.scala
@@ -1,4 +1,4 @@
-package module
+package json_module
 
 import play.api.libs.json._
 import play.api.libs.functional.syntax._
